// implementation/client/side-agent/src/generated/JobState_pb.ts
// JobState.proto — Standard execution states for the Universal Agent Protocol (UAP).

// @generated by protoc-gen-es v1.10.1 with parameter "target=ts,import_extension=.js"
// @generated from file JobState.proto (package uap.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { proto3 } from "@bufbuild/protobuf";

/**
 * JobState represents the state of a test job session.
 *
 * @generated from enum uap.v1.JobState
 */
export enum JobState {
  /**
   * @generated from enum value: JOB_STATE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: JOB_STATE_ACKNOWLEDGED = 1;
   */
  ACKNOWLEDGED = 1,

  /**
   * @generated from enum value: JOB_STATE_INVALID = 2;
   */
  INVALID = 2,

  /**
   * @generated from enum value: JOB_STATE_RUNNING = 3;
   */
  RUNNING = 3,

  /**
   * @generated from enum value: JOB_STATE_COMPLETED = 4;
   */
  COMPLETED = 4,

  /**
   * @generated from enum value: JOB_STATE_FAILED = 5;
   */
  FAILED = 5,
}
// Retrieve enum metadata with: proto3.getEnumType(JobState)
proto3.util.setEnumType(JobState, "uap.v1.JobState", [
  { no: 0, name: "JOB_STATE_UNSPECIFIED" },
  { no: 1, name: "JOB_STATE_ACKNOWLEDGED" },
  { no: 2, name: "JOB_STATE_INVALID" },
  { no: 3, name: "JOB_STATE_RUNNING" },
  { no: 4, name: "JOB_STATE_COMPLETED" },
  { no: 5, name: "JOB_STATE_FAILED" },
]);